Installing Virtual Environment before using it
I used the `pip install virtualenv` command to install the virtual environment on windows based system. I tried running the commmand on my linux terminal for Windows but it did not execute. I then decided to run the same command on Windows Command Prompt and it was able to execute.